Moved camera classes into a camera folder.
