+++ src/vision/cameras/BasicCam.cpp
